requirements.txt
pandas==1.1.5
urllib3==1.23
pyyaml==3.13
requests==2.26.0
numpy==1.16.6
kafka-python==2.0.2
async-timeout==3.0
attrs==17.3.0
aiohttp==3.6.2
pytz==2019.3
psutil==5.8.0
pydantic[dotenv]==1.9.1